Fix passaggio parametri con tipi coerenti
Le entita' vengono salvate come riferimenti in World, e non come
puntatori, visto che una entita' nullptr non ha senso

https://github.com/ranghetto/Qovid/issues/33

+++ src/AI/Patrol.cpp
@@ -1,7 +1,7 @@
 #include "Patrol.h"
 #include "../Controllers/Simulation.h"
 
-Patrol::Patrol(Actor &actor, QVector<QVector2D> waypoints, float waitTime)
+Patrol::Patrol(Actor &actor, QVector<QVector2D> waypoints, uint waitTime)
     : actor_(actor), waypoints_(waypoints), currentWaypointIndex_(0),
       waitTime_(waitTime), waitCounter_(0), waiting_(false) {}
